mutant_app/requirements.txt
Werkzeug==2.0.3
psycopg2-binary==2.9.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
pydantic==2.5.1
pydantic_core==2.14.3
uvicorn==0.24.0.post1
fastapi==0.104.1